Refactores some of the column, row and line code and moves SAVEP
To make room for the longer PID and RC tuning parameter lists the
save and page line had to be moved down. While doing this I
refactored some of the code relating to rows, columns and lines.

Defined a macro for SAVEP_ROW and used it for all code referring
to the SAVEP line. Renamed SAVEP macro to SAVEP_POS as it is
more descriptive. Change some of the config page navigation
code to make it more responsive to changes in page layout.

Tried to use macros instead of hard coded offsets in the code
for text positions. Added macros for title column and column 1, 2
and 3 that can be used instead of the confusing ROLL_T, ROLL_p etc.