Adds new functions for writing MSP requests
Adds four new MSP request functions: mspWriteRequest(), mspWrite8(),
mspWrite16() and mspWriteChecksum(). mspWriteRequest()  takes MSP command and
txDataSize as parameters and writes the MSP header and txSize to the
serial. If txDataSize == 0 the request will be automatically
terminated with the checksum. The request data can be written with
mspWrite8() and mspWrite16() which automatically keeps track of the
checksum. mspWriteChecksum() writes the checksum to the serial and
terminates the request.