Adding support for a generic alarming system
As related to #129, there are many types of alarms the flight controller
may wish to report that aren't necessarily understood by all MWOSD.
This change allows for more flexible alarm reporting.

This code defines a new MSP command "MSP_ALARMS" as 242.  The message
structure is just one a byte for a level and then the alarm message to
display.

The following levels are defined:

* 0 OK
* 1 Warning
* 2 Error
* 3 Critical

This code takes over the "armed/disarmed" line somewhat.  I'm not sure
placement is exactly correct (perhaps could afford longer alert messages
if we could just hijack the whole line).  If there's any reported alarm
from the flight controller, that will be displayed instead of arm
state.  The idea is that an alarm will likely prevent the arming, so
it's good to know why.

Warnings will blink.  Errors and critical alarms will be solid.