AUTOCAM behavior improvement and more stable reset
1. AUTOCAM was recording the camera type in a local static variable,
which caused MAX7456Setup() to be called TWICE at startup. Fixed this
by declaring the camera type as global detectedCamType that is updated
in the MAX7456Setup().

2. RUSHDUINO-Arduino (Nano) based hardware platform started to
experience hardware RESET failures. Fixed this by adding extra 40msec
delay after RESET is pulled high.
  May be applied to the non-RUSHDUINO case also?

3. MAX7456Setup() now executes MAX7456DISABLE before MAX7456HWRESET so
that MAX7456SELECT is high when the MAX gets out of RESET period.

4. Pin mode setup is now declared as MAX7456SETHWPORTS, and executed
only once at the beginning of the setup().

5. Tiding of RUSHDUINO macros for increased readability.