Mavlink wind direction change of operation
Display wind indicator as direction in which wind is heading rather than
direction cming from.
It is more intuitive for FPV and shows better impact on aircraft. e.g.
crosswind for landing.

This is different to other OSD methods. If this is required it can be
enabled in config.h:
#define MAV_WIND_DIR_REVERSE